requirements.txt
Flask==3.1.0
torch==2.6.0
torchvision==0.21.0
Pillow==11.0.0